Relax grammar constraints on what constitutes a valid symbol identifier.
Essentially if it's not a delimiting character, it should
be treated as a symbol character if other primitive data rules
fail to match its token (e.g. numbers, booleans, etc).

Octothorpe (#) and quoting characters ([,'`]) are the only exceptional
(non-delimiting) chararacters; they may not be used to begin an
identifier name.